[IMP] sale: redirect payment links to the customer portal
The customer experience of paying through the payment link was not good.
They had no idea what they were paying for other than the SO reference
and amount. There was no information on whether the current payment
confirmed the quotation or anything. Now, with the link, the client can
access the SO and choose to either pay the amount defined in the link or
the amount that would confirm the order.

task-3087715