[IMP] website_{profile|forum|slides}: edit profile in a popup
Instead of loading a new page for editing the profile, the profile is now
edited in a popup.

To simplify the code, the "edit" link next to the "biography" that allowed the
user to edit it directly in the profile page, is now also replaced by the edit
profile popup (biography field being automatically focused).

Technical notes:
- we have converted the profile edition in OWL instead of using a simple
bootstrap modal because the wysiwyg component was not compatible with bootstrap
modal (ex.: when opening the chatGPT plugin from the wysiwyg field of the
profile bootstrap modal, the prompt field was readonly/not focusable)
- We remove the "edit_button_url_param" and "url_param" that were used to
transmit the forum_id/channel_id to the page /profile/edit so that when saving
on /profile/user/save, the server could redirect to the page
/profile/user with the right parameters. As we have converted the profile
edition to OWL and used it in the same page, the page is refreshed with the
same parameters after saving the profile through ajax call.
- We keep the save controller end point as operation for the current user are
done in sudo so it cannot be replaced by simple orm call from the client.

Task-3975540