[IMP] stock_barcode: show line description in barcode
To adapt for the changes in odoo/enterprise#68629, the
`description_field on `stock.move.line` is set to echo the same value on
its corresponding stock move. Also the default (redundant) value of the
description field which is the name of the product (or code + name)
which was set on fulfilling procurements or on changing product and
picking type is no longer set and left empty instead.

Task-4126824