[IMP] ir.qweb: New `t-arg-*` for `t-call` & `.translate` directive part
The directives `t-attf-*`, `t-argf-*` and `t-valuef-*` can be marked as
as translatable with `.translate`. Translations do not contain python
expressions but only placeholders `{{0}}`, `{{1}}`... This avoids
introducing errors made during the translations. If a new expression is
added `{{unwanted code}}` or a non-existent placeholder, it will be
replaced by `None`.

The `t-arg-*` directive allows to parameterize t-calls and improve the
readability of templates. The t-arg-* directive corresponds to a t-set
defined in the scope of the template rendered by t-call.

`t-call="xmlid" t-arg-varname="python expression"`
`t-call="xmlid" t-argf-varname="string {{expression}}"`
`t-call="xmlid" t-argf.translate-varname="translated string {{expression}}"`