[FIX] microsoft_calendar: avoid recurring editing outlook for other users
Issue: Recurring events created from Outlook should not be able to be
edited from Odoo, this is because the changes made are going to be
sending unwanted emails to the attendees.

Steps to reproduce:
1. Connect to Outlook and Sync for user A.
2. Create a recurring event from Outlook.
3. Connect into a user B with access to the event and edit all the
events from the series.
4. Run the scheduled action to sync user A.
5. Many emails are going to be sent to the attendees from Outlook.

Solution: With the current implementation, we are going to avoid editing
for the users connected to Outlook, however any user with the access to
that event in Odoo and that it has not synced with Outlook is going to
be able to edit the event. To avoid this, we should check if the event
organizer is still synced, if so we assume that the event is also still
synced and we avoid the modifications.

opw-4374503