[IMP] website_sale,*: cart_summary offcanvas on mobile
*:payment_demo,website_event_sale,website_sale,website_sale_loyalty,
website_sale_slides

The summary is taking too much space on small viewports in the checkout.

The goal is to introduce an offcanvas on mobile which contains the
summary and the total table. This also includes small practical and
design revisions.

- Continue & Offcanvas trigger button sticky at the bottom of the page.
- Reviewed spacing
- Items count on top of the image to save horizontal space
- Reordered columns in the code to not rely on order classes
- Move the `o_website_sale_checkout` on the body to apply flex layout on
the `main`
- ! Renders the `express_checkout` button in the button container to
avoid the CSS ugly workaround
- ! Remove the `position-absolute` ugly  workaround for the CTA in the
checkout

task-4594964