[IMP] project: display tasks to plan in calendar side panel
Before this commit, the tasks unplanned are not visible in the calendar
view and so the user has to change the view to be able to see the
unplanned tasks.

This commit adds the tasks to plan in the calendar sidebar if the action
loaded is the one called when the user clicks on a project to see the
tasks of that project. By default, 20 tasks are displayed and a load more
button is displayed if there are more than 20 tasks to plan in the project
to let the user to view more tasks in that side panel.
The unplanned tasks also takes into account filters added in the search
view except if the filters used the planned dates in that case the
domain containing the planned dates is just ignored.

task-4645814