[IMP] project: drag and drop to plan tasks in calendar view
Before this commit, the unplanned tasks are recently visible in the
calendar side panel but the user can only see them and open the form
view when he clicks on them.

This commit allows the user to drag and drop those tasks in the calendar
cells to automatically plan the tasks.
If the scale is in month and year, the task will be planed for the whole
day, otherwise the task will be planned on the time slot selected or the
whole day if the cell is the "column header" (the cell above the time
slots) of the calendar view.

task-4645814