[IMP] stock, *: remove stock.package_level
As packages are now always handled on the move lines, we don't need the
`stock.package_level` model anymore. This means we're no longer
separating moves with/without packages.

To add an entire package to a picking, you can now use the
`action_add_entire_packs` method or the related wizard.

Task-4314600