[IMP] account: display partners no valid sdd mandate
When users try to register one or multiple payments with 'SEPA Direct
Debit' method, and one or more partners don't have a valid SEPA mandate,
a banner alerts the user and mention the names of the concerned
partners. This makes it easier for the user because he directly sees the
partners for which he has to set a SEPA mandate.
This commit add the function called by the <x2many_buttons> widget,
allowing the user to click on the link that redirects to the partner
form.

task-4575537