[FIX] {{sale_}purchase_}stock: grant forecast report access to sales users
Task: 4640948

Currently, a sales user with no stock permissions can see the little
forecast icon and, upon clicking on it, a link to the forecast report.
However, trying to access the report results in an "Access Error"
message.

The forecast report is important for sales users. This commit will allow
them to access the report, but without the option to edit stock pickings
(reserve/unreserve buttons).