[IMP] account: Ease receipt UI/UX
Problem
---------
Currently, receipts are well implemented but highly hidden in the depth
of Odoo. Users are required to activate the option to see and create
receipts.

Objective
---------
Make it easier for users to create and use receipts.

Solution
---------
- Create a new JS widget that mixes the `dynamic_selection` and the
radio widgets to only display possible move types on the move form:
    - From `in_invoice` to `in_receipt` and inversely;
    - From `out_invoice` to `out_receipt` and inversely.
And remove the old views as those are not necessary anymore.
- Add filters to the list view to easily retreive the receipts.
- Remove the two different security groups; receipts are now always
available.
- Add company default receipt taxes that will be used by default on
receipt (over product and account default taxes) to, once again, ease
the user experience. Those default taxes are not available from the UI
because we don't want to cumbersome the settings any more; instead, they
shall be set up in the localization company templates.

task-4677472