[IMP] website: adapt and enable tours
This commit adapts the existing tour website_style_edition. It ensures
that the tours function correctly work with the new data attributes and
structure introduced in the new html_builder. It also enables the tours
that were previously disabled due to the changes in the website
structure.

This commit also removes the focus blur tour and adapt the utils_css.js