[IMP] account,*: Tax Label
Implements a computed field "Tax Label" which holds the value
of the invoice_label with a fallback on the name if empty.

Also replace all the usage of "tax.invoice_label or tax.name"
by this new field.

The aim of this change is to bring more flexibility in the way
we choose what to display in reports, and this new computation
allows to do so without the need to override X reports to achieve
that.