[FIX] pos_hr: allow price control when not explicitly restricted
Before this commit, non-manager cashiers were unable to modify product
prices even when price control was not enabled, which was not the
intended behavior.

Additionally, when `pos_hr` was not installed, the `get_cashier` method
returned the user object with a `role` attribute. However, when
`pos_hr` was installed, the returned cashier used `_role`, leading to
inconsistencies and broken functionality depending on the module's
presence.

opw-4726098

X-original-commit: d28e20ff3c2b366c7c78c5b0a6907f01a2a5732e