[IMP] event: add slot calendar overlay
Purpose
=======
Ease the slots creation in batch which is frustrating as
it's raising a Validation error each time the user tries to
create a slot outside of the event time range.

Specification
=============
Greying out the days outside of the event time range to make
them unselectable. This greatly reduces the chances of having
a Validation error as the users can now visually see the event
time range on the calendar.

As only the days are greyed out and not the hours, a Validation error
can still happen if the user tries to create a slot on the start or end
date of the event outside of its range. For example if the event starts
at 10AM on that day but the user tries to create a slot starting before that.

Task-4743999