[FIX] marketing_card: use preview card in default body
Currently the default mailing body uses the campaign id as a placeholder
for the slug of the individual cards that will be sent.

However this can lead to users getting the wrong card, or an error, when
trying to visit the URL from the mailing preview or when sending it
to themselves using the "test" action.

We now generate an archived card to match the preview image, as was
already done when clicking "preview" from the campaign. So that it behaves
in the preview as it would in the sent email.

task-4247003