[IMP] mail: add notification log on channel and thread rename
Purpose of this PR:
Users can track when a channel and thread is renamed and by whom.

This commit posts a system notification in the channel log
when its name is changed. The message includes the user
who performed the rename and displays the new channel
or thread name.

task-4780916