[FIX] website_event: prevent error when search string has multiple spaces
Currently an error occurs when a user searches for events on the website.

Steps to Reproduce:

 - Install the website_event module.
 - Go to the website and search a string that contains multiple spaces.

NotImplementedError

This error occurs because search.split(' ') returns a list containing empty
strings when the input includes multiple consecutive spaces. The issue
occurs after merging this commit [2]. When the opt(_optimize_like_str)
with [('address_search', 'ilike', '')] (as self ) with event.event model we got
an != operator from here due to we don't have value(search_term) but
we have result. This changes [3] raise an NotImplemented since the != operator
not supported.

This commit ensures that if an empty string is passed to search_in_address,
 it returns the ID in an empty domain to prevent the error.

[1] https://github.com/odoo/odoo/blob/f52d5aff1afa99fa0e8953545b09ad1f7c813b2f/odoo/orm/domains.py

[2] https://github.com/odoo/odoo/commit/92301a5b300dec1ddfca44dc35318b83d67c56fa

[3 ] https://github.com/odoo/odoo/commit/92301a5b300dec1ddfca44dc35318b83d67c56fa#diff-90e43e2412e3476af2026b2726a51ce6050371753243825268e3a21df4fca940R405

sentry-6600544271