[IMP] _*: change cta button if pos_restaurant_appointment is installed
_* = website, website_sale

Before this commit, the CTA button on the website configurator did not
change based on the presence of the `pos_restaurant_appointment` module.

After this commit, if the `pos_restaurant_appointment` module is
installed and the user completes the website setup using the
configurator (instead of skipping and starting from scratch), the
configurator updates the CTA button text to `Book a Table` and links it
to the `/appointment` page.

This POS CTA button will have the highest priority, overriding other
CTA buttons such as `Next Event` or `Schedule an Appointment`.

Additionally, to avoid redundancy, the `Appointment` menu item is
removed from the website when `pos_restaurant_appointment` is installed
and the configurator flow is used.

If the user skips the configurator and starts from scratch, the CTA
button remains unchanged and behaves as it did previously.

Steps to reproduce:

- Install the Point of Sale module.
- Go to Point of Sale > Configuration > Settings.
- Click on Create New Shop.
- Enter the name of the shop and check the `Is a Bar/Restaurant` option.
- In the POS Interface section, enable the `Bookings` checkbox.
- Select an Appointment Type and Save the configuration.
- Go to the Website app.
- Navigate to Configuration > Settings > New Website.
- Fill in the website details and click Create.
- Click `Lets do it` to go through the configurator (do not click `Skip
and start from scratch`).
- Observe that the CTA is set to `Book a Table`, linking to /appointment
and the Appointment menu item is removed.

task-4678541