[FIX] stock: add quantity field dependency lot gen
Steps to Reproduce :

1. Install MRP.
2. Create a Bill of Materials (BOM) and add a component with lot tracking.
3. Create and confirm a Manufacturing Order (MO) with BOM.
4. Navigate to the shop floor and click on Generate Serials/Lots.
5. The error appears when trying to generate the lot .

Issue :

Error raised when try to Generate Serials/Lots during the manufacturing
process.

TypeError: unsupported operand type(s) for //: 'NoneType' and 'int'

Causes:

This error occurs because the quantity field is undefined when
Generate Serials/Lots is triggered, the quantity field is not being passed in
the context, and its value is undefined, which is leading to a division by none.

Solution:

To fix this, the quantity field is explicitly added to the view with
invisible="1" in the XML, ensuring the field is always present and
Its value is correctly passed during the generation of the lot .

opw-4763600